Benchmark 1: main
  Time (mean ± σ):      9.103 s ±  0.044 s    [User: 7.109 s, System: 1.990 s]
  Range (min … max):    9.046 s …  9.157 s    10 runs
 
Benchmark 2: optimizer
  Time (mean ± σ):      4.607 s ±  0.011 s    [User: 4.066 s, System: 0.532 s]
  Range (min … max):    4.594 s …  4.627 s    10 runs
 
Summary
  optimizer ran
    1.98 ± 0.01 times faster than main
    
    where optimizer is this branch